+++ sql/core/src/test/scala/org/apache/spark/sql/sources/CreateTableAsSelectSuite.scala
@@ -17,43 +17,38 @@
 
 package org.apache.spark.sql.sources
 
-import java.io.{File, IOException}
+import java.io.File
 
-import org.scalatest.BeforeAndAfter
+import org.scalatest.BeforeAndAfterAll
 
-import org.apache.spark.sql.AnalysisException
-import org.apache.spark.sql.catalyst.parser.ParseException
-import org.apache.spark.sql.test.SharedSQLContext
+import org.apache.spark.sql.catalyst.util
 import org.apache.spark.util.Utils
 
-class CreateTableAsSelectSuite extends DataSourceTest with SharedSQLContext with BeforeAndAfter {
-  protected override lazy val sql = caseInsensitiveContext.sql _
-  private var path: File = null
+class CreateTableAsSelectSuite extends DataSourceTest with BeforeAndAfterAll {
+
+  import caseInsensisitiveContext._
+
+  var path: File = null
 
   override def beforeAll(): Unit = {
-    super.beforeAll()
-    path = Utils.createTempDir()
+    path = util.getTempFilePath("jsonCTAS").getCanonicalFile
     val rdd = sparkContext.parallelize((1 to 10).map(i => s"""{"a":$i, "b":"str${i}"}"""))
-    caseInsensitiveContext.read.json(rdd).createOrReplaceTempView("jt")
+    jsonRDD(rdd).registerTempTable("jt")
   }
 
   override def afterAll(): Unit = {
-    try {
-      caseInsensitiveContext.dropTempTable("jt")
-    } finally {
-      super.afterAll()
-    }
+    dropTempTable("jt")
   }
 
   after {
-    Utils.deleteRecursively(path)
+    if (path.exists()) Utils.deleteRecursively(path)
   }
 
   test("CREATE TEMPORARY TABLE AS SELECT") {
     sql(
       s"""
         |CREATE TEMPORARY TABLE jsonTable
-        |USING json
+        |USING org.apache.spark.sql.json.DefaultSource
         |OPTIONS (
         |  path '${path.toString}'
         |) AS
@@ -64,37 +59,14 @@ class CreateTableAsSelectSuite extends DataSourceTest with SharedSQLContext with
       sql("SELECT a, b FROM jsonTable"),
       sql("SELECT a, b FROM jt").collect())
 
-    caseInsensitiveContext.dropTempTable("jsonTable")
-  }
-
-  test("CREATE TEMPORARY TABLE AS SELECT based on the file without write permission") {
-    val childPath = new File(path.toString, "child")
-    path.mkdir()
-    childPath.createNewFile()
-    path.setWritable(false)
-
-    val e = intercept[IOException] {
-      sql(
-        s"""
-           |CREATE TEMPORARY TABLE jsonTable
-           |USING json
-           |OPTIONS (
-           |  path '${path.toString}'
-           |) AS
-           |SELECT a, b FROM jt
-        """.stripMargin)
-      sql("SELECT a, b FROM jsonTable").collect()
-    }
-    assert(e.getMessage().contains("Unable to clear output directory"))
-
-    path.setWritable(true)
+    dropTempTable("jsonTable")
   }
 
   test("create a table, drop it and create another one with the same name") {
     sql(
       s"""
         |CREATE TEMPORARY TABLE jsonTable
-        |USING json
+        |USING org.apache.spark.sql.json.DefaultSource
         |OPTIONS (
         |  path '${path.toString}'
         |) AS
@@ -105,107 +77,71 @@ class CreateTableAsSelectSuite extends DataSourceTest with SharedSQLContext with
       sql("SELECT a, b FROM jsonTable"),
       sql("SELECT a, b FROM jt").collect())
 
-    val message = intercept[ParseException]{
+    dropTempTable("jsonTable")
+
+    val message = intercept[RuntimeException]{
       sql(
         s"""
-        |CREATE TEMPORARY TABLE IF NOT EXISTS jsonTable
-        |USING json
+        |CREATE TEMPORARY TABLE jsonTable
+        |USING org.apache.spark.sql.json.DefaultSource
         |OPTIONS (
         |  path '${path.toString}'
         |) AS
         |SELECT a * 4 FROM jt
       """.stripMargin)
     }.getMessage
-    assert(message.toLowerCase.contains("operation not allowed"))
+    assert(
+      message.contains(s"path ${path.toString} already exists."),
+      "CREATE TEMPORARY TABLE IF NOT EXISTS should not be allowed.")
 
-    // Overwrite the temporary table.
-    sql(
-      s"""
-        |CREATE TEMPORARY TABLE jsonTable
-        |USING json
-        |OPTIONS (
-        |  path '${path.toString}'
-        |) AS
-        |SELECT a * 4 FROM jt
-      """.stripMargin)
-    checkAnswer(
-      sql("SELECT * FROM jsonTable"),
-      sql("SELECT a * 4 FROM jt").collect())
-
-    caseInsensitiveContext.dropTempTable("jsonTable")
-    // Explicitly delete the data.
+    // Explicitly delete it.
     if (path.exists()) Utils.deleteRecursively(path)
 
     sql(
       s"""
         |CREATE TEMPORARY TABLE jsonTable
-        |USING json
+        |USING org.apache.spark.sql.json.DefaultSource
         |OPTIONS (
         |  path '${path.toString}'
         |) AS
-        |SELECT b FROM jt
+        |SELECT a * 4 FROM jt
       """.stripMargin)
 
     checkAnswer(
       sql("SELECT * FROM jsonTable"),
-      sql("SELECT b FROM jt").collect())
+      sql("SELECT a * 4 FROM jt").collect())
 
-    caseInsensitiveContext.dropTempTable("jsonTable")
+    dropTempTable("jsonTable")
   }
 
   test("CREATE TEMPORARY TABLE AS SELECT with IF NOT EXISTS is not allowed") {
-    val message = intercept[ParseException]{
+    val message = intercept[DDLException]{
       sql(
         s"""
         |CREATE TEMPORARY TABLE IF NOT EXISTS jsonTable
-        |USING json
+        |USING org.apache.spark.sql.json.DefaultSource
         |OPTIONS (
         |  path '${path.toString}'
         |) AS
         |SELECT b FROM jt
       """.stripMargin)
     }.getMessage
-    assert(message.toLowerCase.contains("operation not allowed"))
+    assert(
+      message.contains("a CREATE TEMPORARY TABLE statement does not allow IF NOT EXISTS clause."),
+      "CREATE TEMPORARY TABLE IF NOT EXISTS should not be allowed.")
   }
 
   test("a CTAS statement with column definitions is not allowed") {
-    intercept[AnalysisException]{
+    intercept[DDLException]{
       sql(
         s"""
         |CREATE TEMPORARY TABLE jsonTable (a int, b string)
-        |USING json
+        |USING org.apache.spark.sql.json.DefaultSource
         |OPTIONS (
         |  path '${path.toString}'
         |) AS
         |SELECT a, b FROM jt
       """.stripMargin)
     }
-  }
-
-  test("it is not allowed to write to a table while querying it.") {
-    sql(
-      s"""
-        |CREATE TEMPORARY TABLE jsonTable
-        |USING json
-        |OPTIONS (
-        |  path '${path.toString}'
-        |) AS
-        |SELECT a, b FROM jt
-      """.stripMargin)
-
-    val message = intercept[AnalysisException] {
-      sql(
-        s"""
-        |CREATE TEMPORARY TABLE jsonTable
-        |USING json
-        |OPTIONS (
-        |  path '${path.toString}'
-        |) AS
-        |SELECT a, b FROM jsonTable
-      """.stripMargin)
-    }.getMessage
-    assert(
-      message.contains("Cannot overwrite table "),
-      "Writing to a table while querying it should not be allowed.")
   }
 }
